requirements.txt
# Librairies de base
tqdm>=4.65.0
python-dotenv>=1.0.0

# Embeddings (sentence-transformers local, CPU)
sentence-transformers>=2.2.2

# RAG et vectorisation
faiss-cpu>=1.7.4
bm25s>=0.3.2
PyStemmer>=2.2.0.3
huggingface_hub>=0.23.0
flashrank>=0.2.10

# Knowledge graph (Neo4j)
neo4j>=5.0.0

# LLM (Cerebras Cloud API — inference ultra-rapide, gratuit)
cerebras-cloud-sdk>=1.0.0

# UI
streamlit>=1.28.0

# Evaluation
pytest>=7.4.0
openai>=1.30.0
ragas>=0.3.0

# Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright>=1.40.0

# Data I/O
jsonlines>=3.1.0
numpy>=1.24.0
